refactor(core): merge @uppy/store-default into core via ./store-default export
