+++ packages/@uppy/core/package.json
@@ -38,11 +38,11 @@
     "./css/style.min.css": "./dist/style.min.css",
     "./css/style.css": "./dist/style.css",
     "./css/style.scss": "./src/style.scss",
+    "./store-default": "./lib/store/index.js",
     "./package.json": "./package.json"
   },
   "dependencies": {
     "@transloadit/prettier-bytes": "^1.1.0",
-    "@uppy/store-default": "workspace:^",
     "@uppy/utils": "workspace:^",
     "lodash": "^4.18.1",
     "mime-match": "^1.0.2",
