+++ packages/@uppy/core/src/store/index.test.ts
