+++ packages/@uppy/core/tsconfig.build.json
@@ -8,9 +8,6 @@
   "exclude": ["./src/**/*.test.ts"],
   "references": [
     {
-      "path": "../store-default/tsconfig.build.json"
-    },
-    {
       "path": "../utils/tsconfig.build.json"
     }
   ]
