+++ packages/@uppy/core/tsconfig.json
@@ -8,9 +8,6 @@
   "include": ["./package.json", "./src/**/*.*"],
   "references": [
     {
-      "path": "../store-default/tsconfig.build.json"
-    },
-    {
       "path": "../utils/tsconfig.build.json"
     }
   ]
