--- packages/@uppy/store-default/.npmignore
@@ -1,1 +1,0 @@
-tsconfig.*
