--- packages/@uppy/store-default/CHANGELOG.md
@@ -1,102 +1,0 @@
-# @uppy/store-default
-
-## 5.0.0
-
-### Major Changes
-
-- c5b51f6: ### Export maps for all packages
-
-  All packages now have export maps. This is a breaking change in two cases:
-
-  1. The css imports have changed from `@uppy[package]/dist/styles.min.css` to `@uppy[package]/css/styles.min.css`
-  2. You were importing something that wasn't exported from the root, for instance `@uppy/core/lib/foo.js`. You can now only import things we explicitly exported.
-
-  #### Changed imports for `@uppy/react`, `@uppy/vue`, and `@uppy/svelte`
-
-  Some components, like Dashboard, require a peer dependency to work but since all components were exported from a single file you were forced to install all peer dependencies. Even if you never imported, for instance, the status bar component.
-
-  Every component that requires a peer dependency has now been moved to a subpath, such as `@uppy/react/dashboard`, so you only need to install the peer dependencies you need.
-
-  **Example for `@uppy/react`:**
-
-  **Before:**
-
-  ```javascript
-  import { Dashboard, StatusBar } from "@uppy/react";
-  ```
-
-  **Now:**
-
-  ```javascript
-  import Dashboard from "@uppy/react/dashboard";
-  import StatusBar from "@uppy/react/status-bar";
-  ```
-
-## 4.3.2
-
-### Patch Changes
-
-- 1b1a9e3: Define "files" in package.json
-
-## 4.3.0
-
-### Minor Changes
-
-- 0c24c5a: Use TypeScript compiler instead of Babel
-
-## 4.1.2
-
-Released: 2024-12-05
-Included in: Uppy v4.8.0
-
-- @uppy/audio,@uppy/aws-s3,@uppy/box,@uppy/companion-client,@uppy/compressor,@uppy/core,@uppy/dashboard,@uppy/drag-drop,@uppy/drop-target,@uppy/dropbox,@uppy/facebook,@uppy/file-input,@uppy/form,@uppy/golden-retriever,@uppy/google-drive,@uppy/google-photos,@uppy/image-editor,@uppy/informer,@uppy/instagram,@uppy/locales,@uppy/onedrive,@uppy/progress-bar,@uppy/provider-views,@uppy/react,@uppy/remote-sources,@uppy/screen-capture,@uppy/status-bar,@uppy/store-default,@uppy/thumbnail-generator,@uppy/transloadit,@uppy/tus,@uppy/unsplash,@uppy/url,@uppy/utils,@uppy/vue,@uppy/webcam,@uppy/xhr-upload,@uppy/zoom: cleanup tsconfig (Mikael Finstad / #5520)
-
-## 4.1.1
-
-Released: 2024-10-31
-Included in: Uppy v4.6.0
-
-- @uppy/aws-s3,@uppy/box,@uppy/companion-client,@uppy/core,@uppy/dashboard,@uppy/drag-drop,@uppy/dropbox,@uppy/facebook,@uppy/file-input,@uppy/form,@uppy/golden-retriever,@uppy/google-drive,@uppy/google-photos,@uppy/image-editor,@uppy/informer,@uppy/instagram,@uppy/locales,@uppy/onedrive,@uppy/progress-bar,@uppy/provider-views,@uppy/react-native,@uppy/react,@uppy/redux-dev-tools,@uppy/screen-capture,@uppy/status-bar,@uppy/store-default,@uppy/store-redux,@uppy/svelte,@uppy/thumbnail-generator,@uppy/transloadit,@uppy/tus,@uppy/unsplash,@uppy/url,@uppy/utils,@uppy/vue,@uppy/webcam,@uppy/xhr-upload,@uppy/zoom: Fix links (Anthony Veaudry / #5492)
-
-## 4.1.0
-
-Released: 2024-07-30
-Included in: Uppy v4.1.0
-
-- @uppy/core,@uppy/store-default: export `Store` type (Merlijn Vos / #5373)
-
-## 3.1.0
-
-Released: 2023-11-24
-Included in: Uppy v3.20.0
-
-- @uppy/store-default: refactor to typescript (Antoine du Hamel / #4785)
-
-## 3.0.2
-
-Released: 2022-09-25
-Included in: Uppy v3.1.0
-
-- @uppy/audio,@uppy/aws-s3-multipart,@uppy/aws-s3,@uppy/box,@uppy/companion-client,@uppy/companion,@uppy/compressor,@uppy/core,@uppy/dashboard,@uppy/drag-drop,@uppy/drop-target,@uppy/dropbox,@uppy/facebook,@uppy/file-input,@uppy/form,@uppy/golden-retriever,@uppy/google-drive,@uppy/image-editor,@uppy/informer,@uppy/instagram,@uppy/locales,@uppy/onedrive,@uppy/progress-bar,@uppy/provider-views,@uppy/react,@uppy/redux-dev-tools,@uppy/remote-sources,@uppy/screen-capture,@uppy/status-bar,@uppy/store-default,@uppy/store-redux,@uppy/svelte,@uppy/thumbnail-generator,@uppy/transloadit,@uppy/tus,@uppy/unsplash,@uppy/url,@uppy/utils,@uppy/vue,@uppy/webcam,@uppy/xhr-upload,@uppy/zoom: add missing entries to changelog for individual packages (Antoine du Hamel / #4092)
-
-## 3.0.0
-
-Released: 2022-08-22
-Included in: Uppy v3.0.0
-
-- Switch to ESM
-
-## 3.0.0-beta.2
-
-Released: 2022-08-03
-Included in: Uppy v3.0.0-beta.4
-
-- @uppy/store-default: export the class, don't expose `.callbacks` (Antoine du Hamel / #3928)
-
-## 2.1.0
-
-Released: 2022-05-30
-Included in: Uppy v2.11.0
-
-- @uppy/angular,@uppy/audio,@uppy/aws-s3-multipart,@uppy/aws-s3,@uppy/box,@uppy/core,@uppy/dashboard,@uppy/drag-drop,@uppy/dropbox,@uppy/facebook,@uppy/file-input,@uppy/form,@uppy/golden-retriever,@uppy/google-drive,@uppy/image-editor,@uppy/informer,@uppy/instagram,@uppy/onedrive,@uppy/progress-bar,@uppy/react,@uppy/redux-dev-tools,@uppy/robodog,@uppy/screen-capture,@uppy/status-bar,@uppy/store-default,@uppy/store-redux,@uppy/thumbnail-generator,@uppy/transloadit,@uppy/tus,@uppy/unsplash,@uppy/url,@uppy/vue,@uppy/webcam,@uppy/xhr-upload,@uppy/zoom: doc: update bundler recommendation (Antoine du Hamel / #3763)
-- @uppy/store-default: refactor to ESM (Antoine du Hamel / #3746)
