--- packages/@uppy/store-default/package.json
@@ -1,43 +1,0 @@
-{
-  "name": "@uppy/store-default",
-  "description": "The default simple object-based store for Uppy.",
-  "version": "5.0.0",
-  "license": "MIT",
-  "main": "lib/index.js",
-  "type": "module",
-  "sideEffects": false,
-  "scripts": {
-    "build": "tsc --build tsconfig.build.json",
-    "typecheck": "tsc --build",
-    "test": "vitest run --environment=jsdom --silent='passed-only'"
-  },
-  "keywords": [
-    "file uploader",
-    "uppy",
-    "uppy-store"
-  ],
-  "homepage": "https://uppy.io",
-  "bugs": {
-    "url": "https://github.com/transloadit/uppy/issues"
-  },
-  "devDependencies": {
-    "@types/node": "^20.19.0",
-    "jsdom": "^29.1.1",
-    "typescript": "^5.8.3",
-    "vitest": "^4.1.6"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/transloadit/uppy.git"
-  },
-  "exports": {
-    ".": "./lib/index.js",
-    "./package.json": "./package.json"
-  },
-  "files": [
-    "src",
-    "lib",
-    "dist",
-    "CHANGELOG.md"
-  ]
-}
