--- packages/@uppy/store-default/tsconfig.build.json
@@ -1,10 +1,0 @@
-{
-  "extends": "../../../tsconfig.shared",
-  "compilerOptions": {
-    "outDir": "./lib",
-    "rootDir": "./src"
-  },
-  "include": ["./src/**/*.*"],
-  "exclude": ["./src/**/*.test.ts"],
-  "references": []
-}
