--- packages/@uppy/store-default/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "extends": "../../../tsconfig.shared",
-  "compilerOptions": {
-    "emitDeclarationOnly": false,
-    "noEmit": true,
-    "types": ["preact", "node"]
-  },
-  "include": ["./package.json", "./src/**/*.*"],
-  "references": []
-}
