+++ packages/uppy/package.json
@@ -61,7 +61,6 @@
     "@uppy/remote-sources": "workspace:*",
     "@uppy/screen-capture": "workspace:*",
     "@uppy/status-bar": "workspace:*",
-    "@uppy/store-default": "workspace:*",
     "@uppy/thumbnail-generator": "workspace:*",
     "@uppy/transloadit": "workspace:*",
     "@uppy/tus": "workspace:*",
