+++ packages/uppy/tsconfig.build.json
@@ -68,9 +68,6 @@
       "path": "../@uppy/screen-capture/tsconfig.build.json"
     },
     {
-      "path": "../@uppy/store-default/tsconfig.build.json"
-    },
-    {
       "path": "../@uppy/thumbnail-generator/tsconfig.build.json"
     },
     {
