+++ packages/uppy/tsconfig.json
@@ -67,9 +67,6 @@
       "path": "../@uppy/screen-capture/tsconfig.build.json"
     },
     {
-      "path": "../@uppy/store-default/tsconfig.build.json"
-    },
-    {
       "path": "../@uppy/thumbnail-generator/tsconfig.build.json"
     },
     {
